I moved the postgres container's logic to IntegrationTestBase
