+++ src/test/java/ru/o0000q/ratealbumservice/integration/IntegrationTestBase.java
@@ -1,5 +1,6 @@
 package ru.o0000q.ratealbumservice.integration;
 
+import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.DynamicPropertyRegistry;
@@ -9,21 +10,29 @@ import org.springframework.transaction.annotation.Transactional;
 import org.testcontainers.containers.PostgreSQLContainer;
 import ru.o0000q.ratealbumservice.integration.annotation.IT;
 
-//@Sql({
-//    "classpath:sql/fill_database.sql"
-//})
+
 @IT
 @Sql({
         "classpath:sql/fill_database.sql"
 })
 public abstract class IntegrationTestBase {
-//    private static final PostgreSQLContainer<?> container = new PostgreSQLContainer<>("postgres");
-//    @BeforeAll
-//    static void runContainer() {
-//        container.start();
-//    }
-//    @DynamicPropertySource
-//    static void postgresProperties(DynamicPropertyRegistry registry) {
-//        registry.add("spring.datasource.url", container::getJdbcUrl);
-//    }
+    static PostgreSQLContainer<?> postgres = new PostgreSQLContainer<>(
+            "postgres:latest"
+    );
+    @BeforeAll
+    static void beforeAll() {
+        postgres.start();
+    }
+
+    @AfterAll
+    static void afterAll() {
+        postgres.stop();
+    }
+
+    @DynamicPropertySource
+    static void configureProperties(DynamicPropertyRegistry registry) {
+        registry.add("spring.datasource.url", postgres::getJdbcUrl);
+        registry.add("spring.datasource.username", postgres::getUsername);
+        registry.add("spring.datasource.password", postgres::getPassword);
+    }
 }
